backend/pyproject.toml
[project]
name = "marcel"
version = "0.9.2"
description = "Default template for PDM package"
authors = [
    {name = "Jan Trienes", email = "jan.trienes@uni-marburg.de"},
]
dependencies = [
    "haystack-ai==2.14.2",
    "sentence-transformers==4.1.0",
    "fastapi[standard]==0.115.12",
    "uvicorn[standard]==0.34.2",
    "transformers[sentencepiece,torch]==4.52.2",
    "nltk==3.9.1",
    "sqlalchemy[asyncio]>=2.0.40",
    "alembic==1.16.1",
    "w3lib==2.3.1",
    "PyMySQL==1.1.1",
    "pyjwt>=2.10.1",
    "argon2-cffi>=23.1.0",
    "asyncmy>=0.2.10"
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

# This adds the cpu-only wheels for torch to the solving index. Important is `include_packages` and `exclude_packages`
# to ensure fast solving. Earlier we implemented a cpu-group and locked this for linux platform.
# For reference (these do not work):
# https://github.com/pdm-project/pdm/issues/2696
# https://github.com/pytorch/pytorch/issues/136275#issuecomment-2469957219
[[tool.pdm.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"
include_packages = ["torch"]
exclude_packages = ["*"]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
# Needed when mixing async and non-async tests
# https://github.com/nedbat/coveragepy/issues/1082#issuecomment-979462293
concurrency = ["thread", "greenlet"]

[tool.ruff.format]
exclude = ["*.ipynb"]


[tool.pdm]
# This is set to true so that package becomes importable
distribution = true

[tool.pdm.scripts]
pre_dev.cmd = "bash scripts/prestart.sh"
pre_dev.env_file = '../.env'
dev.cmd = "fastapi dev src/marcel/main.py"
dev.env_file = "../.env"
pre_prod.cmd = "bash scripts/prestart.sh"
pre_prod.env_file = '../.env'
prod.cmd = "uvicorn marcel.main:app --host 0.0.0.0 --port 9000"
prod.env_file = "../.env"
format-code = "ruff format"
format-imports = "ruff check --select I --fix"
format = {composite = ["format-imports", "format-code"]}
lint = "ruff check"
lint-fix = "ruff check --fix"
test.cmd = "pytest --cov-report html --cov=marcel --cov-branch"
test.env_file = '../.env.test'

[dependency-groups]
dev = [
    "pytest==8.3.5",
    "pytest-mock==3.14.0",
    "pytest-cov==6.1.1",
    "ipykernel==6.29.5",
    "ruff==0.11.7",
    "pytest-asyncio>=0.25.3",
    "ipywidgets>=8.1.5",
    "pandas>=2.2.3",
    "openpyxl>=3.1.5",
    "aiosqlite>=0.21.0",
    "asgi-lifespan==2.*",
]
